// src/components/button/Enums.ts
export enum EButtonPosition {
    RIGHT = 'right',
    LEFT = 'left',
    CENTER = 'center'
}

export enum EButtonColor {
    LIGHT_COLOR = 'light-color',
    DARK_COLOR = 'dark-color',
    BRAND_COLOR = 'brand-color',
    DARKEN_COLOR = 'darken-color'
}

export enum EButtonTextPosition {
    LEFT = 'left',
    RIGHT = 'right'
}